Equipment Screen is in game. Player Can use Items to heal outside of battle
BUG FIXES:
Bug made the player walk through everything in some zones. Fixed
Memory leaks with TriggerMatch.cpp not releasing skills when game hardclosed.
Bug in Use Items, with new Equipment being implemented had to change useItem to check if an item was a consumable before checking if the cursor id was correct